Remove old Create WAN action, replace fully with new one
